requirements.txt
Flask
pandas
requests
ta
websocket-client
numpy
selenium
sseclient-py
webdriver-manager
requests